pyproject.toml
[project]
name = "eia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.1.0",
    "loguru>=0.7.3",
    "mlflow>=3.8.1",
    "nbformat>=5.10.4",
    "pandas>=2.3.3",
    "pandas-stubs~=2.3.3",
    "plotly>=6.6.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "ruff>=0.14.10",
]
